# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mwango Brain API Documentation",
        "description": "Documentação das APIs RESTful do sistema Mwango Brain - Gestão de Eventos, Leads e Contactos (Laravel 13 & Neon PostgreSQL)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local / Dev"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Autenticação de utilizador",
                "operationId": "login",
                "responses": {
                    "200": {
                        "description": "Login efetuado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
...
        "/api/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Obter contacto por ID",
                "operationId": "getContactById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto encontrado"
                    },
                    "404": {
                        "description": "Contacto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Atualizar contacto existente",
                "operationId": "updateContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto atualizado"
                    },
                    "404": {
                        "description": "Contacto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Eliminar contacto",
                "operationId": "deleteContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto eliminado"
                    }
                }
            }
        },
...
        "/api/follow-ups/{id}/complete": {
            "patch": {
                "tags": [
                    "FollowUps"
                ],
                "summary": "Concluir follow-up",
                "operationId": "completeFollowUp",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow-up marcado como concluído"
                    }
                }
            }
        },
        "/api/follow-ups/{id}": {
            "put": {
                "tags": [
                    "FollowUps"
                ],
                "summary": "Atualizar estado do follow-up",
                "operationId": "updateFollowUp",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow-up atualizado"
                    }
                }
            }
        },
...
    },
...
}